router 설정 파일명 변경
