+++ src/components/HeaderComponent.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import router from '@/router'
+import router from '@/router/router'
 
 const userid: String = '루챌'
 </script>
