+++ src/router/router.ts
@@ -1,0 +1,9 @@
+import { createRouter, createWebHistory } from 'vue-router'
+import routes from '@/router/routes'
+
+const router = createRouter({
+  history: createWebHistory(import.meta.env.BASE_URL),
+  routes: routes
+})
+
+export default router
